Fixed some visual bugs on main water and deleted reset password button.

+++ app/src/main/java/com/rainyteam/views/MainWaterActivity.kt
@@ -180,7 +180,7 @@ class MainWaterActivity : AppCompatActivity(), CoroutineScope {
                 ContextCompat.getDrawable(getApplicationContext(), R.drawable.bottle_90)
             )
         }
-        textWaterPercent!!.text = actualUserWater.toString() + "%"
+        textWaterPercent!!.text = actualUserWater.toInt() .toString() + "%"
 
     }
